Enhancement/optimizations 2 (#123)
* added log

* added prod bucket host to next.config.js

* fix initial loading

* removed mb from holiday message

* fixed email

* Now showing real logos

* testing mobile skeleton grid

* fix

* fix

* various adjustments

* testing 280px

* 200px

* testing 240px

* testing notifications

* use client

* removed deviceId hook call

* testing

* added push notifications logic

* fix base url

* device id

* fix device id

* fix

* sw try catch

* testing

* fix

* testing

* notifications preferences

* not firing push when disabling

* fix

* fix fetch

* refactor

* fix polling

* fix polling

* schedule now passing subscribed to schedule row

* testing mobile behaviour

* fix

* testing mobile fix

* fix

* fix device id generation

* fix program id passing

* deleting duplicated push

* added pwa for iOS

* fix iOS

* testing visual adjustments

* replaced 100vh with 100dvh

* added register and login modal

* login modal restyling

* fix login route

* fix login modal

* restyling

* added new routes

* fixed email get not found

* login button improvement

* testing id 0

* testing fix

* testing

* fix loading state

* fix skeleton grid

* fix white screen while loading

* acceder fix

* changed typography variant

* applied MUI to login modal

* fix

* optimization

* fix user name

* fixed email search

* removed import

* testing

* fix name and style

* testing

* fix and added logs

* fix login legacy

* fix isBackoffice

* fix again

* logs

* login legacy final fix

* first stage

* fix linter

* fix users email url

* second stage

* updated changelog

* removed backoffice login

* added backoffice user menu option

* is admin verification

* added profile page

* updated changelog

* added reset password feature

* header improvement

* updated changelog

* added users backoffice

* updated changelog and menuItems

* fix build

* icons

* icons unification MUI

* fix email endpoint

* applied visual improvements

* reverted code step change

* sessions and ux improvements

* adjustments

* several visual adjustments

* improved user button

* improvements to user menu

* updated changelog

* added user role to user backoffice

* integrated subsctiptions with users

* updated changelog

* fixes

* rounded corners for grid element

* fixed subscriptions not being generated

* fix

* visual adjustments

* day button padding

* fix push

* testing

* removed useless request to backend

* optimization home client

* fix

* hardcoding tuesday

* fix

* optimized profile page

* optimized subscriptions page

* fix subscriptions page

* fixed subscriptions title

* updated changelog

* fixed rest of the week schedules

* fix server date time

* fixed program blocks moving

* fixed animations

* only show panelists on programs longer than 1 hour

* conditioned panelists in mobile

* fix tooltips and redirect to login

* several fixes

* fix

* fix tooltip theme

* tooltip style

* fixed tooltip

* fix

* fix

* fix schedules backoffice

* defensive fixes

* logs

* fixed backoffice reloading

* fix

* updated changelog

* changelog date
