+++ src/app/legal/page.tsx
@@ -3,32 +3,63 @@ export const dynamic = 'force-dynamic';
 
 import { ChannelWithSchedules } from '@/types/channel';
 import HomeClient from '@/components/HomeClient';
+import { ClientWrapper } from '@/components/ClientWrapper';
+import { getBuenosAiresDayOfWeek } from '@/utils/date';
 
+interface InitialData {
+  holiday: boolean;
+  todaySchedules: ChannelWithSchedules[];
+  weekSchedules: ChannelWithSchedules[];
+}
 
 export default async function Page() {
-  // Calcula el día de la semana en inglés en minúsculas
-  const today = new Date()
-    .toLocaleString('en-US', { weekday: 'long' })
-    .toLowerCase();
+  // Get today's day of week using Buenos Aires time
+  const today = getBuenosAiresDayOfWeek();
 
   const url = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3000';
 
-  // Fetch inicial (ISR cada 60s), con fallback seguro en caso de fallo
-  let initialData: ChannelWithSchedules[] = [];
+  // Fetch both today's and week's schedules in parallel
+  let todaySchedules: ChannelWithSchedules[] = [];
+  let weekSchedules: ChannelWithSchedules[] = [];
+
   try {
-    const res = await fetch(
-      `${url}/channels/with-schedules?day=${today}`,
-      { next: { revalidate: 60 } }
-    );
-    if (res.ok) {
-      initialData = await res.json();
+    const [todayRes, weekRes] = await Promise.all([
+      fetch(
+        `${url}/channels/with-schedules?day=${today}`,
+        { next: { revalidate: 60 } }
+      ),
+      fetch(
+        `${url}/channels/with-schedules`,
+        { next: { revalidate: 60 } }
+      )
+    ]);
+
+    if (todayRes.ok) {
+      todaySchedules = await todayRes.json();
     } else {
-      console.warn('Fetch failed with status', res.status);
+      console.warn('Today schedules fetch failed with status', todayRes.status);
+    }
+
+    if (weekRes.ok) {
+      weekSchedules = await weekRes.json();
+    } else {
+      console.warn('Week schedules fetch failed with status', weekRes.status);
     }
   } catch (err) {
     console.warn('Fetch error during build/runtime:', err);
   }
 
+  // Prepare the initial data with the correct structure
+  const initialData: InitialData = {
+    holiday: false, // Legal page doesn't need holiday info
+    todaySchedules,
+    weekSchedules
+  };
+
   // Renderiza componente cliente con datos pre-cargados
-  return <HomeClient initialData={initialData} />;
+  return (
+    <ClientWrapper>
+      <HomeClient initialData={initialData} />
+    </ClientWrapper>
+  );
 }
