+++ src/components/ProgramBlock.tsx
@@ -66,15 +66,14 @@ export const ProgramBlock: React.FC<Props> = ({
   const { pixelsPerMinute } = useLayoutValues();
   const { mode } = useThemeContext();
   const [isMobile, setIsMobile] = useState(false);
-  const [openTooltip, setOpenTooltip] = useState(false);
   const bellRef = useRef<HTMLButtonElement>(null);
   const [isOn, setIsOn] = useState(subscribed);
-
-  // Refs para controlar delay de apertura y cierre
-  const openTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const closeTimeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
   const { openVideo, openPlaylist } = useYouTubePlayer();
   const { subscribeAndRegister } = usePush();
+  const [openTooltip, setOpenTooltip] = useState(false);
+  const openTimeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const closeTimeoutRef = useRef<NodeJS.Timeout | null>(null);
 
   useEffect(() => {
     setIsOn(subscribed);
@@ -86,49 +85,6 @@ export const ProgramBlock: React.FC<Props> = ({
       setIsMobile(/Mobi|Android/i.test(navigator.userAgent));
     }
   }, []);
-
-  // Limpieza de timeouts al desmontar
-  useEffect(() => {
-    return () => {
-      if (openTimeoutRef.current) clearTimeout(openTimeoutRef.current);
-      if (closeTimeoutRef.current) clearTimeout(closeTimeoutRef.current);
-    };
-  }, []);
-
-  // Apertura retardada 5s
-  const handleTooltipOpen = () => {
-    if (closeTimeoutRef.current) {
-      clearTimeout(closeTimeoutRef.current);
-      closeTimeoutRef.current = null;
-    }
-    if (!isMobile) {
-      if (openTimeoutRef.current) clearTimeout(openTimeoutRef.current);
-      openTimeoutRef.current = setTimeout(() => {
-        Clarity.setTag('program_name', name);
-        Clarity.event(isLive ? 'open_tooltip_live' : 'open_tooltip_deferred');
-        gaEvent(
-          isLive ? 'open_tooltip_live' : 'open_tooltip_deferred',
-          {
-          category: 'tooltip',
-          program_name: name,
-        });
-        setOpenTooltip(true);
-      }, 500);
-    }
-  };
-
-  // Cierre rápido
-  const handleTooltipClose = () => {
-    if (openTimeoutRef.current) {
-      clearTimeout(openTimeoutRef.current);
-      openTimeoutRef.current = null;
-    }
-    if (!isMobile) {
-      closeTimeoutRef.current = setTimeout(() => {
-        setOpenTooltip(false);
-      }, 100);
-    }
-  };
 
   // Cálculo de posición y tamaño
   const [startHours, startMinutes] = start.split(':').map(Number);
@@ -185,9 +141,14 @@ export const ProgramBlock: React.FC<Props> = ({
       return;
     }
 
-    try {
-      const willSubscribe = !isOn;
+    const prevIsOn = isOn;
+    const willSubscribe = !isOn;
 
+    // Optimistically update UI
+    setIsOn(willSubscribe);
+    setIsLoading(true);
+
+    try {
       if (willSubscribe) {
         // Get push subscription details if needed
         let pushSubscription = null;
@@ -218,18 +179,57 @@ export const ProgramBlock: React.FC<Props> = ({
             headers: { Authorization: `Bearer ${typedSession.accessToken}` },
           }
         );
-        setIsOn(true);
         console.log(`✅ Subscribed to ${name}`);
       } else {
         // Unsubscribe from program
         await api.delete(`/programs/${id}/subscribe`, {
           headers: { Authorization: `Bearer ${typedSession.accessToken}` },
         });
-        setIsOn(false);
         console.log(`🚫 Unsubscribed from ${name}`);
       }
+      // Success: do nothing, UI already updated
     } catch (error) {
+      // Revert UI on error
+      setIsOn(prevIsOn);
+      alert('Error updating subscription. Please try again.');
       console.error('Error updating subscription:', error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  // Limpieza de timeouts al desmontar
+  useEffect(() => {
+    return () => {
+      if (openTimeoutRef.current) clearTimeout(openTimeoutRef.current);
+      if (closeTimeoutRef.current) clearTimeout(closeTimeoutRef.current);
+    };
+  }, []);
+
+  // Apertura retardada 500ms
+  const handleTooltipOpen = () => {
+    if (closeTimeoutRef.current) {
+      clearTimeout(closeTimeoutRef.current);
+      closeTimeoutRef.current = null;
+    }
+    if (!isMobile) {
+      if (openTimeoutRef.current) clearTimeout(openTimeoutRef.current);
+      openTimeoutRef.current = setTimeout(() => {
+        setOpenTooltip(true);
+      }, 500);
+    }
+  };
+
+  // Cierre rápido
+  const handleTooltipClose = () => {
+    if (openTimeoutRef.current) {
+      clearTimeout(openTimeoutRef.current);
+      openTimeoutRef.current = null;
+    }
+    if (!isMobile) {
+      closeTimeoutRef.current = setTimeout(() => {
+        setOpenTooltip(false);
+      }, 100);
     }
   };
 
@@ -237,26 +237,24 @@ export const ProgramBlock: React.FC<Props> = ({
   const tooltipContent = (
     <Box
       sx={{ p: tokens.spacing.sm }}
-      onMouseEnter={handleTooltipOpen}
-      onMouseLeave={handleTooltipClose}
     >
-      <Text variant="subtitle1" fontWeight={tokens.typography.fontWeight.bold} color="white">
+      <Text variant="subtitle1" fontWeight={tokens.typography.fontWeight.bold} sx={{ color: '#fff !important' }}>
         {name}
       </Text>
-      <Text variant="body2" sx={{ mt: tokens.spacing.sm, color: 'rgba(255,255,255,0.9)' }}>
+      <Text variant="body2" sx={{ mt: tokens.spacing.sm, color: 'rgba(255,255,255,0.8) !important' }}>
         {start} - {end}
       </Text>
       {description && (
-        <Text variant="body2" sx={{ mt: tokens.spacing.sm, color: 'rgba(255,255,255,0.9)' }}>
+        <Text variant="body2" sx={{ mt: tokens.spacing.sm, color: 'rgba(255,255,255,0.8) !important' }}>
           {description}
         </Text>
       )}
       {panelists?.length ? (
         <Box sx={{ mt: tokens.spacing.sm }}>
-          <Text variant="body2" fontWeight={tokens.typography.fontWeight.bold} color="white">
+          <Text variant="body2" fontWeight={tokens.typography.fontWeight.bold} sx={{ color: '#fff !important' }}>
             Panelistas:
           </Text>
-          <Text variant="body2" color="rgba(255,255,255,0.9)">
+          <Text variant="body2" sx={{ color: 'rgba(255,255,255,0.8) !important' }}>
             {panelists.map(p => p.name).join(', ')}
           </Text>
         </Box>
@@ -288,8 +286,23 @@ export const ProgramBlock: React.FC<Props> = ({
         aria-label="Notificarme"
         onClick={handleBellClick}
         ref={bellRef}
+        sx={{
+          mt: tokens.spacing.md,
+          color: isLoading ? undefined : (isOn ? '#fff !important' : 'rgba(255,255,255,0.5) !important'),
+        }}
+        disabled={isLoading}
       >
-        {isOn ? <Notifications color="primary" /> : <Notifications color="disabled" />}
+        {isLoading ? (
+          <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: 24, width: 24 }}>
+            <svg width="20" height="20" viewBox="0 0 40 40" style={{ display: 'block' }}>
+              <circle cx="20" cy="20" r="18" stroke="#1976d2" strokeWidth="4" fill="none" strokeDasharray="90" strokeDashoffset="60">
+                <animateTransform attributeName="transform" type="rotate" from="0 20 20" to="360 20 20" dur="1s" repeatCount="indefinite" />
+              </circle>
+            </svg>
+          </Box>
+        ) : (
+          <Notifications sx={{ color: isOn ? '#fff !important' : 'rgba(255,255,255,0.5) !important' }} />
+        )}
       </IconButton>
     </Box>
   );
@@ -300,7 +313,6 @@ export const ProgramBlock: React.FC<Props> = ({
         if (bellRef.current?.contains(event.target as Node)) {
           return;
         }
-
         if (isMobile) setOpenTooltip(false);
       }
     }>
@@ -320,15 +332,18 @@ export const ProgramBlock: React.FC<Props> = ({
           onMouseEnter={handleTooltipOpen}
           onMouseLeave={handleTooltipClose}
           onClick={() => isMobile && setOpenTooltip(!openTooltip)}
-          position="absolute"
-          left={`${offsetPx}px`}
-          width={`${widthPx}px`}
+          style={{
+            position: 'absolute',
+            left: `${offsetPx}px`,
+            width: `${widthPx}px`,
+            height: '100%',
+          }}
           height="100%"
           sx={{
             backgroundColor: alpha(color, isPast ? 0.05 : isLive ? (mode === 'light' ? 0.2 : 0.3) : (mode === 'light' ? 0.1 : 0.15)),
             border: `1px solid ${isPast ? alpha(color, mode === 'light' ? 0.3 : 0.4) : color}`,
             borderRadius: tokens.borderRadius.sm,
-            transition: `all ${tokens.transition.normal} ${tokens.transition.timing}`,
+            transition: `background-color ${tokens.transition.normal} ${tokens.transition.timing}`,
             cursor: 'pointer',
             overflow: 'hidden',
             boxShadow: tokens.boxShadow.sm,
@@ -409,7 +424,7 @@ export const ProgramBlock: React.FC<Props> = ({
                 >
                   {name.toUpperCase()}
                 </Typography>
-                {panelists && panelists.length > 0 && (
+                {panelists && panelists.length > 0 && (!isMobile || (isMobile && widthPx > 120)) && (
                   <Typography
                     variant="caption"
                     sx={{
