+++ src/contexts/LiveStatusContext.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { createContext, useContext, useState, useEffect } from 'react';
 import dayjs from 'dayjs';
